requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Database
asyncpg>=0.28.0
psycopg2-binary>=2.9.0

# LLM and AI
aiohttp>=3.8.0
requests>=2.31.0

# Scientific computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization  
plotly>=5.15.0
folium>=0.14.0

# Data formats
netCDF4>=1.6.0

# Optional: Enhanced features
# jupyter>=1.0.0          # For notebook interface
# streamlit>=1.25.0       # For web interface
# fastapi>=0.100.0        # For API interface
# uvicorn>=0.22.0         # For API server

# Development dependencies (uncomment for development)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0